RF+TST: Move repodate's set_date helper to tests/utils
This will be used in the tests for the check_dates extension too.
